Make deploy script check if .env exist, abort if not

+++ deploy.sh
@@ -1,5 +1,10 @@
 #!/bin/bash
 
+if [ ! -f .env ]; then
+    echo ".env is not found! Aborting!"
+    exit 1
+fi
+
 git pull origin main
 docker compose build
 docker compose down && docker compose up -d
